+++ packages/builder/canvas/src/form/index.tsx
@@ -6,11 +6,14 @@ export function FormBuilderCanvas() {
   const select = useCanvas(({ select }) => select);
   const { ref } = useClipboard();
 
+  const handleSelect = () => select({ selectedBlockIds: null });
+
   return (
     <Canvas
       ref={ref}
       className="min-h-min py-10"
-      onClick={() => select({ selectedBlockIds: null })}
+      onClick={handleSelect}
+      onKeyDown={handleSelect}
     >
       <div className="w-[500px] space-y-2 rounded">
         <FormDisplay />
